Split up individual.ts
Split up individual.ts
